# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AppNexus Proxy API",
        "description": "A rest-ful API to proxy AppNexus.",
        "contact": {
            "email": "friends@niiknow.org"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
        "/proxy/{path}": {
            "get": {
                "tags": [
                    "proxy"
                ],
                "summary": "get proxy",
                "parameters": [
                    {
                        "name": "X-API-Key",
                        "in": "header",
                        "description": "api key",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "path",
                        "in": "path",
                        "description": "path",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "response object"
                    }
                }
            },
            "put": {
                "tags": [
                    "proxy"
                ],
                "summary": "put proxy",
                "parameters": [
                    {
                        "name": "X-API-Key",
                        "in": "header",
                        "description": "api key",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "path",
                        "in": "path",
                        "description": "path",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "response object"
                    }
                }
            },
            "post": {
                "tags": [
                    "proxy"
                ],
                "summary": "post proxy",
                "parameters": [
                    {
                        "name": "X-API-Key",
                        "in": "header",
                        "description": "api key",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "path",
                        "in": "path",
                        "description": "path",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "response object"
                    }
                }
            },
            "delete": {
                "tags": [
                    "proxy"
                ],
                "summary": "delete proxy",
                "operationId": "Api\\Controllers\\ProxyController::index",
                "parameters": [
                    {
                        "name": "X-API-Key",
                        "in": "header",
                        "description": "api key",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "path",
                        "in": "path",
                        "description": "path",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "response object"
                    }
                }
            }
        }
    }
}